requirements/development.txt
-r deployment.txt
coverage==3.7.1
dj-database-url
django-devserver==0.8.0
django-extensions
django-i18n-helper==0.1.1
django-nose
django-template-repl==0.3.0
django-webtest==1.9.1
factory_boy==2.4.1
freezegun==0.3.5
ipdb==0.10.3
mock==1.0.1
nose==1.3.7
nosexcover==1.0.10
path.py==10.3.1
pinocchio==0.3.1
pre-commit
pudb==2016.2
PyVirtualDisplay==0.1.5
PyYAML==3.12
sphinx==1.5.6
sqlparse==0.2.3
spec
transifex-client==0.12.4
unittest-data-provider==1.0.1
virtualenv-tools==1.0
webtest==2.0.27
werkzeug==0.12.2
